+++ test/common/http/utility_test.cc
@@ -457,7 +457,8 @@ TEST(HttpUtility, SendLocalReply) {
 
   EXPECT_CALL(callbacks, encodeHeaders_(_, false));
   EXPECT_CALL(callbacks, encodeData(_, true));
-  Utility::sendLocalReply(false, callbacks, is_reset, Http::Code::PayloadTooLarge, "large", false);
+  Utility::sendLocalReply(false, callbacks, is_reset, Http::Code::PayloadTooLarge, "large",
+                          absl::nullopt, false);
 }
 
 TEST(HttpUtility, SendLocalGrpcReply) {
@@ -468,11 +469,37 @@ TEST(HttpUtility, SendLocalGrpcReply) {
       .WillOnce(Invoke([&](const HeaderMap& headers, bool) -> void {
         EXPECT_STREQ(headers.Status()->value().c_str(), "200");
         EXPECT_NE(headers.GrpcStatus(), nullptr);
-        EXPECT_STREQ(headers.GrpcStatus()->value().c_str(), "2"); // Unknown gRPC error.
+        EXPECT_EQ(headers.GrpcStatus()->value().c_str(),
+                  std::to_string(enumToInt(Grpc::Status::GrpcStatus::Unknown)));
         EXPECT_NE(headers.GrpcMessage(), nullptr);
         EXPECT_STREQ(headers.GrpcMessage()->value().c_str(), "large");
       }));
-  Utility::sendLocalReply(true, callbacks, is_reset, Http::Code::PayloadTooLarge, "large", false);
+  Utility::sendLocalReply(true, callbacks, is_reset, Http::Code::PayloadTooLarge, "large",
+                          absl::nullopt, false);
+}
+
+TEST(HttpUtility, RateLimitedGrpcStatus) {
+  MockStreamDecoderFilterCallbacks callbacks;
+
+  EXPECT_CALL(callbacks, encodeHeaders_(_, true))
+      .WillOnce(Invoke([&](const HeaderMap& headers, bool) -> void {
+        EXPECT_NE(headers.GrpcStatus(), nullptr);
+        EXPECT_EQ(headers.GrpcStatus()->value().c_str(),
+                  std::to_string(enumToInt(Grpc::Status::GrpcStatus::Unavailable)));
+      }));
+  Utility::sendLocalReply(true, callbacks, false, Http::Code::TooManyRequests, "", absl::nullopt,
+                          false);
+
+  EXPECT_CALL(callbacks, encodeHeaders_(_, true))
+      .WillOnce(Invoke([&](const HeaderMap& headers, bool) -> void {
+        EXPECT_NE(headers.GrpcStatus(), nullptr);
+        EXPECT_EQ(headers.GrpcStatus()->value().c_str(),
+                  std::to_string(enumToInt(Grpc::Status::GrpcStatus::ResourceExhausted)));
+      }));
+  Utility::sendLocalReply(
+      true, callbacks, false, Http::Code::TooManyRequests, "",
+      absl::make_optional<Grpc::Status::GrpcStatus>(Grpc::Status::GrpcStatus::ResourceExhausted),
+      false);
 }
 
 TEST(HttpUtility, SendLocalReplyDestroyedEarly) {
@@ -483,7 +510,8 @@ TEST(HttpUtility, SendLocalReplyDestroyedEarly) {
     is_reset = true;
   }));
   EXPECT_CALL(callbacks, encodeData(_, true)).Times(0);
-  Utility::sendLocalReply(false, callbacks, is_reset, Http::Code::PayloadTooLarge, "large", false);
+  Utility::sendLocalReply(false, callbacks, is_reset, Http::Code::PayloadTooLarge, "large",
+                          absl::nullopt, false);
 }
 
 TEST(HttpUtility, SendLocalReplyHeadRequest) {
@@ -494,7 +522,8 @@ TEST(HttpUtility, SendLocalReplyHeadRequest) {
         EXPECT_STREQ(headers.ContentLength()->value().c_str(),
                      fmt::format("{}", strlen("large")).c_str());
       }));
-  Utility::sendLocalReply(false, callbacks, is_reset, Http::Code::PayloadTooLarge, "large", true);
+  Utility::sendLocalReply(false, callbacks, is_reset, Http::Code::PayloadTooLarge, "large",
+                          absl::nullopt, true);
 }
 
 TEST(HttpUtility, TestExtractHostPathFromUri) {